Refactor SwarmFunctions Class
