+++ DSAFinal/src/org/neu/pso/swarmfunctions/SwarmFunctions.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.neu.pso.config;
+package org.neu.pso.swarmfunctions;
 
 /**
  *
